backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
pyyaml
pandas
rapidfuzz
httpx
